task-72: moved interface IMember to interfaces folder
