+++ src/js/interfaces/IMember.ts
